+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Change keyslot for New 3DS key sector keys from 0x11 to 0x43
   * This adds a new Keyslot enum item: `Keyslot.New3DSKeySector`
 * Document more methods in pyctr.crypto.engine
+* Add new `fileio.CloseWrapper` class to provide access to a file object while preventing closing it directly
 
 ## v0.4.1 - July 11, 2020
 * Support NCCH contents with fixed crypto key (zerokey and fixed system key)